requirements.txt
PyGithub==1.55
pyyaml==6.0
requests==2.28.1
pylint==2.15.5
flake8==6.0.0
markdown==3.4.1
python-dotenv==0.21.0
openai==1.3.0 